[Feat] 즐겨찾기 및 댓글 개수 홈뷰의 게시물뷰에 표시

+++ TodaySpace/TodaySpace/Resource/Int+.swift
@@ -1,5 +1,5 @@
 //
-//  Double+.swift
+//  Int+.swift
 //  TodaySpace
 //
 //  Created by 김상규 on 2/28/25.
@@ -7,6 +7,6 @@
 
 import Foundation
 
-extension Double: DecimalString {
+extension Int: DecimalString {
     var decimalValue: NSNumber { return NSNumber(value: self) }
 }
